backend/app/services/inference/generate.py
"""
Image generation service - STUB implementation.
TODO: Replace with actual diffusers inference pipeline.
"""
import time
from pathlib import Path
from typing import Dict, Any, Optional
from PIL import Image, ImageDraw, ImageFont
from app.core.logging import get_logger

logger = get_logger(__name__)


def generate_image(
    prompt: str,
    negative_prompt: Optional[str] = None,
    model_version_id: int = None,
    lora_path: Optional[str] = None,
    steps: int = 50,
    width: int = 512,
    height: int = 512,
    seed: Optional[int] = None,
    output_path: str = None
) -> str:
    """
    Generate image (STUB - placeholder implementation).
    
    Args:
        prompt: Text prompt
        negative_prompt: Negative prompt
        model_version_id: Model version ID
        lora_path: Path to LoRA weights (optional)
        steps: Number of steps
        width: Image width
        height: Image height
        seed: Random seed
        output_path: Path to save generated image
    
    Returns:
        Path to generated image
    """
    logger.info(
        "generation_started",
        prompt=prompt[:50],  # Log first 50 chars
        steps=steps,
        width=width,
        height=height
    )
    
    # STUB: Simulate generation
    time.sleep(0.5)  # Simulate work
    
    # STUB: Create placeholder image
    output_file = Path(output_path) if output_path else Path(f"output_{int(time.time())}.png")
    output_file.parent.mkdir(parents=True, exist_ok=True)
    
    # Create a simple placeholder image with text
    img = Image.new('RGB', (width, height), color='white')
    draw = ImageDraw.Draw(img)
    
    # Try to use a font, fallback to default
    try:
        font = ImageFont.truetype("arial.ttf", 24)
    except:
        font = ImageFont.load_default()
    
    # Draw prompt text
    text = f"STUB\nPrompt: {prompt[:30]}...\nSteps: {steps}\nSize: {width}x{height}"
    if seed:
        text += f"\nSeed: {seed}"
    
    # Center text
    bbox = draw.textbbox((0, 0), text, font=font)
    text_width = bbox[2] - bbox[0]
    text_height = bbox[3] - bbox[1]
    position = ((width - text_width) // 2, (height - text_height) // 2)
    
    draw.text(position, text, fill='black', font=font)
    
    # Add border
    draw.rectangle([0, 0, width-1, height-1], outline='gray', width=2)
    
    img.save(output_file)
    
    logger.info("generation_completed", output_path=str(output_file))
    
    return str(output_file)


def generate_thumbnail(image_path: str, thumbnail_path: str, size: tuple = (256, 256)) -> str:
    """Generate thumbnail from image."""
    img = Image.open(image_path)
    img.thumbnail(size, Image.Resampling.LANCZOS)
    img.save(thumbnail_path)
    return thumbnail_path


# TODO: Implement real generation using diffusers
# Example structure:
# def generate_image_real(prompt, lora_path, ...):
#     from diffusers import StableDiffusionPipeline
#     from peft import PeftModel
#     # Load base model
#     pipe = StableDiffusionPipeline.from_pretrained(base_model)
#     # Load LoRA
#     pipe.unet = PeftModel.from_pretrained(pipe.unet, lora_path)
#     # Generate
#     image = pipe(prompt, ...).images[0]
#     return image
