backend/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Celery
celery==5.3.4
redis==5.0.1

# S3 / MinIO
boto3==1.29.7
minio==7.2.0

# Image processing
Pillow==10.1.0
imagehash==4.3.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Logging
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
